MULE-8990: State param encoding in OAuth support not compatible with certain services like Concur
Concur (and others new connectors) complains about getting characters like '<'  '>' in the URL parameters, claiming it is a dangerous request.
Currently, We used it in the constant variable EVENT_STATE_TEMPLATE.
Devkit needs a mechanism to replace this character when they generate code.
With this fix, Devkit will able to implement the getEventStateTemplateSuffix method to override the respective character.
If the implemented method returns null, mule.oauth.custom.suffix system property will be evaluated, and only if it wasn't set, so the default ">>" prefix will be used.
"<<" Prefix will be removed because It's not used.